refactor(query): extract shared SQL datasource package
Addresses review feedback on PR #776: the Athena formatter, response
types, response parsing, and LIMIT-clamping logic were near-verbatim
copies of the ClickHouse equivalents.

Introduce internal/query/sql with the common building blocks shared by
SQL-style Grafana datasources:
- QueryResponse/Column row-oriented result types
- FormatTable/FormatWideTable table rendering
- ParseResponse (refId "A" -> rows), product-named for errors
- EnforceLimit core clamp/append, with a per-dialect bail predicate

ClickHouse and Athena now own only their dialect-specific concerns:
request construction, schema discovery (system tables vs the Resource
API), and their LIMIT bail rules. The query codec collapses the two
identical *QueryResponse switch arms into one *sql.QueryResponse case.
Net -148 lines; behavior preserved (dialect EnforceLimit tests unchanged).
