+++ internal/datasources/query/codecs.go
@@ -15,6 +15,7 @@ import (
 	"github.com/grafana/gcx/internal/query/loki"
 	"github.com/grafana/gcx/internal/query/prometheus"
 	"github.com/grafana/gcx/internal/query/pyroscope"
+	querysql "github.com/grafana/gcx/internal/query/sql"
 	"github.com/grafana/gcx/internal/query/tempo"
 )
 
@@ -44,14 +45,12 @@ func (c *queryTableCodec) Encode(w io.Writer, data any) error {
 		return influxdb.FormatQueryTable(w, resp)
 	case *tempo.GetTraceResponse:
 		return tempo.FormatTraceTable(w, resp)
-	case *clickhouse.QueryResponse:
-		return clickhouse.FormatTable(w, resp)
+	case *querysql.QueryResponse:
+		return querysql.FormatTable(w, resp)
 	case []clickhouse.TableInfo:
 		return clickhouse.FormatListTablesTable(w, resp)
 	case []clickhouse.ColumnInfo:
 		return clickhouse.FormatDescribeTableTable(w, resp)
-	case *athena.QueryResponse:
-		return athena.FormatTable(w, resp)
 	case athena.StringList:
 		return athena.FormatStringList(w, resp.Items, resp.Header)
 	case *cloudwatch.QueryResponse:
@@ -83,10 +82,8 @@ func (c *queryWideCodec) Encode(w io.Writer, data any) error {
 		return infinity.FormatTable(w, resp)
 	case *tempo.GetTraceResponse:
 		return tempo.FormatTraceWide(w, resp)
-	case *clickhouse.QueryResponse:
-		return clickhouse.FormatWideTable(w, resp)
-	case *athena.QueryResponse:
-		return athena.FormatWideTable(w, resp)
+	case *querysql.QueryResponse:
+		return querysql.FormatWideTable(w, resp)
 	case athena.StringList:
 		return athena.FormatStringList(w, resp.Items, resp.Header)
 	case *cloudwatch.QueryResponse:
@@ -147,14 +144,12 @@ func (c *queryGraphCodec) Encode(w io.Writer, data any) error {
 		if err != nil {
 			return err
 		}
-	case *clickhouse.QueryResponse:
-		return errors.New("graph output is not supported for ClickHouse queries; use -o table/json/yaml")
+	case *querysql.QueryResponse:
+		return errors.New("graph output is not supported for SQL datasource queries; use -o table/json/yaml")
 	case []clickhouse.TableInfo:
 		return errors.New("graph output is not supported for ClickHouse list-tables; use -o table/json/yaml")
 	case []clickhouse.ColumnInfo:
 		return errors.New("graph output is not supported for ClickHouse describe-table; use -o table/json/yaml")
-	case *athena.QueryResponse:
-		return errors.New("graph output is not supported for Athena queries; use -o table/json/yaml")
 	case athena.StringList:
 		return errors.New("graph output is not supported for Athena discovery; use -o table/json/yaml")
 	default:
