--- internal/query/athena/export_test.go
@@ -1,6 +1,0 @@
-package athena
-
-// ParseResponse exposes parseResponse for testing.
-func ParseResponse(body []byte) (*QueryResponse, error) {
-	return parseResponse(body)
-}
