+++ internal/query/sql/parse_test.go
@@ -1,27 +1,28 @@
-package clickhouse_test
+package sql_test
 
 import (
 	"encoding/json"
 	"testing"
 
-	"github.com/grafana/gcx/internal/query/clickhouse"
+	"github.com/grafana/gcx/internal/query/dataframe"
+	"github.com/grafana/gcx/internal/query/sql"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestParseResponse_SingleFrame(t *testing.T) {
-	raw := clickhouse.GrafanaQueryResponse{
-		Results: map[string]clickhouse.GrafanaResult{
+	raw := dataframe.Response{
+		Results: map[string]dataframe.Result{
 			"A": {
-				Frames: []clickhouse.DataFrame{
+				Frames: []dataframe.Frame{
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{
 								{Name: "id", Type: "number"},
 								{Name: "name", Type: "string"},
 							},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{
 								{float64(1), float64(2)},
 								{"alice", "bob"},
@@ -37,10 +38,10 @@ func TestParseResponse_SingleFrame(t *testing.T) {
 	body, err := json.Marshal(raw)
 	require.NoError(t, err)
 
-	resp, err := clickhouse.ParseResponse(body)
+	resp, err := sql.ParseResponse(body, "clickhouse")
 	require.NoError(t, err)
 
-	assert.Equal(t, []clickhouse.Column{{Name: "id", Type: "number"}, {Name: "name", Type: "string"}}, resp.Columns)
+	assert.Equal(t, []sql.Column{{Name: "id", Type: "number"}, {Name: "name", Type: "string"}}, resp.Columns)
 	require.Len(t, resp.Rows, 2)
 	assert.InDelta(t, 1, resp.Rows[0][0], 0)
 	assert.Equal(t, "alice", resp.Rows[0][1])
@@ -49,23 +50,23 @@ func TestParseResponse_SingleFrame(t *testing.T) {
 }
 
 func TestParseResponse_UsesOnlyFirstFrame(t *testing.T) {
-	raw := clickhouse.GrafanaQueryResponse{
-		Results: map[string]clickhouse.GrafanaResult{
+	raw := dataframe.Response{
+		Results: map[string]dataframe.Result{
 			"A": {
-				Frames: []clickhouse.DataFrame{
+				Frames: []dataframe.Frame{
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{{Name: "v", Type: "number"}},
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{{Name: "v", Type: "number"}},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{{float64(1), float64(2)}},
 						},
 					},
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{{Name: "v", Type: "number"}},
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{{Name: "v", Type: "number"}},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{{float64(3)}},
 						},
 					},
@@ -78,7 +79,7 @@ func TestParseResponse_UsesOnlyFirstFrame(t *testing.T) {
 	body, err := json.Marshal(raw)
 	require.NoError(t, err)
 
-	resp, err := clickhouse.ParseResponse(body)
+	resp, err := sql.ParseResponse(body, "clickhouse")
 	require.NoError(t, err)
 
 	require.Len(t, resp.Rows, 2)
@@ -87,23 +88,23 @@ func TestParseResponse_UsesOnlyFirstFrame(t *testing.T) {
 }
 
 func TestParseResponse_MismatchedFrameSkipped(t *testing.T) {
-	raw := clickhouse.GrafanaQueryResponse{
-		Results: map[string]clickhouse.GrafanaResult{
+	raw := dataframe.Response{
+		Results: map[string]dataframe.Result{
 			"A": {
-				Frames: []clickhouse.DataFrame{
+				Frames: []dataframe.Frame{
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{{Name: "a", Type: "string"}},
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{{Name: "a", Type: "string"}},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{{"x"}},
 						},
 					},
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{{Name: "b", Type: "number"}},
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{{Name: "b", Type: "number"}},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{{float64(99)}},
 						},
 					},
@@ -116,17 +117,17 @@ func TestParseResponse_MismatchedFrameSkipped(t *testing.T) {
 	body, err := json.Marshal(raw)
 	require.NoError(t, err)
 
-	resp, err := clickhouse.ParseResponse(body)
+	resp, err := sql.ParseResponse(body, "clickhouse")
 	require.NoError(t, err)
 
-	assert.Equal(t, []clickhouse.Column{{Name: "a", Type: "string"}}, resp.Columns)
+	assert.Equal(t, []sql.Column{{Name: "a", Type: "string"}}, resp.Columns)
 	require.Len(t, resp.Rows, 1)
 	assert.Equal(t, "x", resp.Rows[0][0])
 }
 
 func TestParseResponse_ErrorInResult(t *testing.T) {
-	raw := clickhouse.GrafanaQueryResponse{
-		Results: map[string]clickhouse.GrafanaResult{
+	raw := dataframe.Response{
+		Results: map[string]dataframe.Result{
 			"A": {
 				Error:       "Code: 62. DB::Exception: Syntax error",
 				ErrorSource: "downstream",
@@ -138,21 +139,21 @@ func TestParseResponse_ErrorInResult(t *testing.T) {
 	body, err := json.Marshal(raw)
 	require.NoError(t, err)
 
-	_, err = clickhouse.ParseResponse(body)
+	_, err = sql.ParseResponse(body, "clickhouse")
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "Syntax error")
 }
 
 func TestParseResponse_EmptyResult(t *testing.T) {
-	raw := clickhouse.GrafanaQueryResponse{
-		Results: map[string]clickhouse.GrafanaResult{
+	raw := dataframe.Response{
+		Results: map[string]dataframe.Result{
 			"A": {
-				Frames: []clickhouse.DataFrame{
+				Frames: []dataframe.Frame{
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{{Name: "x", Type: "string"}},
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{{Name: "x", Type: "string"}},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{{}},
 						},
 					},
@@ -165,23 +166,23 @@ func TestParseResponse_EmptyResult(t *testing.T) {
 	body, err := json.Marshal(raw)
 	require.NoError(t, err)
 
-	resp, err := clickhouse.ParseResponse(body)
+	resp, err := sql.ParseResponse(body, "clickhouse")
 	require.NoError(t, err)
 
-	assert.Equal(t, []clickhouse.Column{{Name: "x", Type: "string"}}, resp.Columns)
+	assert.Equal(t, []sql.Column{{Name: "x", Type: "string"}}, resp.Columns)
 	assert.Empty(t, resp.Rows)
 }
 
 func TestParseResponse_MissingRefID(t *testing.T) {
-	raw := clickhouse.GrafanaQueryResponse{
-		Results: map[string]clickhouse.GrafanaResult{
+	raw := dataframe.Response{
+		Results: map[string]dataframe.Result{
 			"B": {
-				Frames: []clickhouse.DataFrame{
+				Frames: []dataframe.Frame{
 					{
-						Schema: clickhouse.DataFrameSchema{
-							Fields: []clickhouse.DataFrameField{{Name: "v", Type: "number"}},
+						Schema: dataframe.Schema{
+							Fields: []dataframe.Field{{Name: "v", Type: "number"}},
 						},
-						Data: clickhouse.DataFrameData{
+						Data: dataframe.Data{
 							Values: [][]any{{float64(1)}},
 						},
 					},
@@ -194,7 +195,7 @@ func TestParseResponse_MissingRefID(t *testing.T) {
 	body, err := json.Marshal(raw)
 	require.NoError(t, err)
 
-	resp, err := clickhouse.ParseResponse(body)
+	resp, err := sql.ParseResponse(body, "clickhouse")
 	require.NoError(t, err)
 
 	assert.Empty(t, resp.Columns)
